Fix PartitionedAssetTimetable asset handling
This needs to eventually be migrated to SDK too. I don't know how yet...
